Integrate string_jump_state into deserializer callback code
